optee-os: set log level to 2 (INFO) for all platforms
Defines the OP-TEE core log level to be 2 (INFO) for all platforms,
unless overridden in the platform .inc file. Prior to this change was
the log level equal to 1 (ERROR) for all platforms.

The INFO level shows the OP-TEE banner on boot with version, date and
compiler info and also shows which CPUs are initialized. INFO messages
are typically not used during normal operation of OP-TEE, so it should
an issue to have them enabled even if OP-TEE shares its console with
Normal World.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>