optee-os: cache RPMB FAT data to improve performance
Set CFG_RPMB_FS_CACHE_ENTRIES to 48 so that the first 48*256 bytes of
RPMB data are cached in the TEE core heap memory. It is where the File
Allocation Table resides, and therefore we get a performance boost
every time a persistent object is accessed.

As an example, this improves the execution time of "xtest 6007" which
is takes 10.4 seconds before this patch and 4.3 seconds after.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>