optee-os: select best fit allocation algorithm for TEE core heap
Use the best fit allocation algorithm for the OP-TEE core heap in
order to reduce fragmentation. This solves occasional errors with the
PKCS#11 early TA which fails to load on RockPi4 at least (error message:
"Decompression error (-4)"). The performance impact should be marginal
and the memory usage much more efficient.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>