rockpi4: u-boot: do not prevent parallel build
I see no reason to set PARELLEL_BUILD="" in the RockPi4 recipe for
U-Boot. This line is not documented in the commit that introduced it
nor in the recipe file itself. I have succesfully built U-Boot a number
of time on a 36-core machine with this change applied so I think it is
valid. It saves about one minute.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>