optee-ftpm: deploy non-stripped fTPM ELF file for debugging purposes
The optee-ftpm recipe copies the stripped ELF binary of the TA into the
deploy directory, since it is what is ultimately run by OP-TEE. However
it is useful to copy the non-stripped version too for debugging purposes.
The ELF file with symbolic and debugging information is needed to
interpret TA crash dumps using OP-TEE's symbolize.py script.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>